% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{combinations}
\alias{combinations}
\title{combinations}
\usage{
combinations(p)
}
\arguments{
\item{p}{Dimension of covariates.}
}
\value{
A matrix containing all the possible missing patterns. Each row indicates a pattern of missingness. "1" means "observed", 0 means "missing".
}
\description{
Given all the possible patterns of missingness.
}
\examples{
comb = combinations(5)
}
