% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.complete_doc.R
\docType{data}
\name{data.complete}
\alias{data.complete}
\title{Complete dataset}
\description{
Complete dataset
}
\references{
\url{https://sites.google.com/site/weijiangstat/}{homepage}
}
\author{
Wei Jiang \email{wei.jiang@polytechnique.edu}
}
\keyword{complete}
\keyword{dataset}
