% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{imputeEllP}
\alias{imputeEllP}
\title{Function for imputing single point for linear regression model}
\usage{
imputeEllP(point, Sigma.inv)
}
\arguments{
\item{point}{A single observation containing missing values.}

\item{Sigma.inv}{Inverse of estimated \eqn{\Sigma}{\Sigma}.}
}
\value{
Imputed observation.
}
\description{
Function for imputing single point for linear regression model
}
