% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_reg.R
\name{log_reg}
\alias{log_reg}
\title{log_reg}
\usage{
log_reg(y, x, beta, iflog = TRUE)
}
\arguments{
\item{y}{Response value (0 or 1).}

\item{x}{Covariate vector of dimension \eqn{p \times 1}{p*1}.}

\item{beta}{Estimated parameter of logistic regression model.}

\item{iflog}{If TRUE, log_reg calculate the log-likelihood; else likelihood.}
}
\value{
Likelihood or log-likelihood.
}
\description{
Caculate the likelihood or log-likelihood for one observation of logistic regression model .
}
\examples{
res = log_reg(1,c(1,2,3),c(1,-1,1))
}
