% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/louis_lr_saem.R
\name{louis_lr_saem}
\alias{louis_lr_saem}
\title{louis_lr_saem}
\usage{
louis_lr_saem(beta, mu, Sigma, Y, X.obs, pos_var, rindic, whichcolXmissing,
  mc.size)
}
\arguments{
\item{beta}{Estimated parameter of logistic regression model.}

\item{mu}{Estimated parameter \eqn{\mu}.}

\item{Sigma}{Estimated parameter \eqn{\Sigma}.}

\item{Y}{Response vector \eqn{N \times 1}{N * 1}}

\item{X.obs}{Design matrix with missingness \eqn{N \times p}{N * p}}

\item{pos_var}{Index of selected covariates.}

\item{rindic}{Missing pattern of X.obs. If a component in X.obs is missing, the corresponding position in rindic is 1; else 0.}

\item{whichcolXmissing}{The column index in covariate containing at least one missing observation.}

\item{mc.size}{Monte Carlo sampling size.}
}
\value{
Variance of estimated \eqn{\beta}.
}
\description{
Used in main function miss.saem. Caculate the variance of estimated parameters for logistic regression model with missing data, using Monte Carlo version of Louis formula.
}
