% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_no_na.R
\name{mean_no_na}
\alias{mean_no_na}
\title{Mean}
\usage{
mean_no_na(x)
}
\arguments{
\item{x}{a numerique vector}
}
\value{
the mean
}
\description{
This computes the mean without na
}
\examples{
mean_no_na(c(4,5))
}
