% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misaem_doc.R
\docType{package}
\name{misaem}
\alias{misaem}
\alias{SAEM}
\alias{for}
\alias{logistic}
\alias{regression}
\alias{with}
\alias{missing}
\alias{data}
\alias{misaem-package}
\title{misaem : SAEM for logistic regression with missing data}
\description{
Detailled description of your package
}
\author{
Wei Jiang <wei.jiang@polytechnique.edu>
}
