% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss.lm.control.R
\name{miss.lm.control}
\alias{miss.lm.control}
\title{Auxiliary for Controlling Fitting}
\usage{
miss.lm.control(maxruns = 500, tol_em = 1e-07, print_iter = TRUE)
}
\arguments{
\item{maxruns}{maximum number of iterations. The default is maxruns = 500.}

\item{tol_em}{the tolerance to stop EM. The default is tol_em = 1e-4.}

\item{print_iter}{logical indicating if output should be produced for each iteration.}
}
\value{
A list with components named as the arguments.
}
\description{
Auxiliary function for \code{\link{miss.lm}} fitting. Typically only used internally by \code{\link{miss.lm.fit}}.
}
\examples{
## For examples see example(miss.lm)
}
