% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_fun.R
\name{my_fun}
\alias{my_fun}
\title{Title}
\usage{
my_fun(x, y)
}
\arguments{
\item{x}{my param}

\item{y}{my second param}
}
\value{
this return x+y
}
\description{
Description
}
\examples{
my_fun(4,5)
my_fun(5,9)
}
