% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.miss.glm.R
\name{print.miss.glm}
\alias{print.miss.glm}
\title{Print miss.glm}
\usage{
\method{print}{miss.glm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class "\code{miss.glm}", usually, a result of a call to \code{\link{miss.glm}}.}

\item{digits}{minimal number of significant digits.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print results for class \code{miss.glm}.
}
\examples{
## For examples see example(miss.glm)
}
