% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.miss.lm.R
\name{print.miss.lm}
\alias{print.miss.lm}
\title{Print miss.lm}
\usage{
\method{print}{miss.lm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class "\code{miss.lm}", usually, a result of a call to \code{\link{miss.lm}}.}

\item{digits}{minimal number of significant digits.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print results for class \code{miss.lm}.
}
\examples{
## For examples see example(miss.lm)
}
