% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.miss.glm.old.R
\name{print.summary.miss.glm.old}
\alias{print.summary.miss.glm.old}
\title{Print Summary of miss.glm}
\usage{
\method{print}{summary.miss.glm.old}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class "\code{summary.miss.glm}", usually, a result of a call to \code{\link{summary.miss.glm}}.}

\item{digits}{minimal number of significant digits.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print results for class \code{summary.miss.glm}.
}
\examples{
## For examples see example(miss.glm)
}
