% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.miss.glm.old.R
\name{summary.miss.glm.old}
\alias{summary.miss.glm.old}
\title{Summarizing Fits for miss.glm}
\usage{
\method{summary}{miss.glm.old}(object, ...)
}
\arguments{
\item{object}{an object of class "\code{miss.glm}", usually, a result of a call to \code{\link{miss.glm}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class "\code{summary.miss.glm}", a list with following components:
\item{coefficients}{The matrix of coefficients and standard errors}
\item{loglikelihood}{Observed log-likelihood.}
\item{call}{the component from \code{object}.}
\item{formula}{the component from \code{object}.}
}
\description{
Summary for class \code{miss.glm}.
}
\examples{
## For examples see example(miss.glm)
}
