% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.miss.lm.R
\name{summary.miss.lm}
\alias{summary.miss.lm}
\title{Summarizing Fits for miss.lm}
\usage{
\method{summary}{miss.lm}(object, ...)
}
\arguments{
\item{object}{an object of class "\code{miss.lm}", usually, a result of a call to \code{\link{miss.lm}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class "\code{summary.miss.lm}", a list with following components:
\item{coefficients}{The matrix of coefficients and standard errors.}
\item{loglikelihood}{Observed log-likelihood.}
\item{call}{the component from \code{object}.}
\item{formula}{the component from \code{object}.}
}
\description{
Summary for class \code{miss.lm}.
}
\examples{
## For examples see example(miss.lm)
}
