% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.complete_doc.R
\docType{data}
\name{X.complete}
\alias{X.complete}
\title{Covriables in complete dataset}
\description{
Covriables in complete dataset
}
\references{
\url{https://sites.google.com/site/weijiangstat/}{homepage}
}
\author{
Wei Jiang \email{wei.jiang@polytechnique.edu}
}
\keyword{complete}
\keyword{covriables}
\keyword{dataset}
\keyword{in}
