% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.obs.MAR_doc.R
\docType{data}
\name{X.obs.MAR}
\alias{X.obs.MAR}
\title{Covriables in dataset with missingness (missing at random)}
\description{
Covriables in dataset with missingness (missing at random)
}
\references{
\url{https://sites.google.com/site/weijiangstat/}{homepage}
}
\author{
Wei Jiang \email{wei.jiang@polytechnique.edu}
}
\keyword{(missing}
\keyword{at}
\keyword{covriables}
\keyword{dataset}
\keyword{in}
\keyword{missingness}
\keyword{random)}
\keyword{the}
\keyword{with}
