% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_doc.R
\docType{data}
\name{y}
\alias{y}
\title{Responce vairbale in dataset}
\description{
Responce vairbale in dataset
}
\references{
\url{https://sites.google.com/site/weijiangstat/}{homepage}
}
\author{
Wei Jiang \email{wei.jiang@polytechnique.edu}
}
\keyword{Responce}
\keyword{The}
\keyword{a}
\keyword{and}
\keyword{any}
\keyword{binary}
\keyword{contain}
\keyword{data}
\keyword{dataset.}
\keyword{doesn't}
\keyword{in}
\keyword{is}
\keyword{missing}
\keyword{responce}
\keyword{vairbale}
\keyword{variable.}
